// client/stocklinx app/src/redux/productStatus/constant.ts
export enum productStatusConst {
  FETCH_PRODUCTSTATUSES_REQUEST = "FETCH_PRODUCTSTATUSES_REQUEST",
  FETCH_PRODUCTSTATUSES_SUCCESS = "FETCH_PRODUCTSTATUSES_SUCCESS",
  FETCH_PRODUCTSTATUSES_FAILURE = "FETCH_PRODUCTSTATUSES_FAILURE",
  FETCH_PRODUCTSTATUS_REQUEST = "FETCH_PRODUCTSTATUS_REQUEST",
  FETCH_PRODUCTSTATUS_SUCCESS = "FETCH_PRODUCTSTATUS_SUCCESS",
  FETCH_PRODUCTSTATUS_FAILURE = "FETCH_PRODUCTSTATUS_FAILURE",
  CREATE_PRODUCTSTATUS_REQUEST = "CREATE_PRODUCTSTATUS_REQUEST",
  CREATE_PRODUCTSTATUS_SUCCESS = "CREATE_PRODUCTSTATUS_SUCCESS",
  CREATE_PRODUCTSTATUS_FAILURE = "CREATE_PRODUCTSTATUS_FAILURE",
  UPDATE_PRODUCTSTATUS_REQUEST = "UPDATE_PRODUCTSTATUS_REQUEST",
  UPDATE_PRODUCTSTATUS_SUCCESS = "UPDATE_PRODUCTSTATUS_SUCCESS",
  UPDATE_PRODUCTSTATUS_FAILURE = "UPDATE_PRODUCTSTATUS_FAILURE",
  REMOVE_PRODUCTSTATUS_REQUEST = "REMOVE_PRODUCTSTATUS_REQUEST",
  REMOVE_PRODUCTSTATUS_SUCCESS = "REMOVE_PRODUCTSTATUS_SUCCESS",
  REMOVE_PRODUCTSTATUS_FAILURE = "REMOVE_PRODUCTSTATUS_FAILURE",
  SET_PRODUCTSTATUS = "SET_PRODUCTSTATUS",
  CLEAR_PRODUCTSTATUS = "CLEAR_PRODUCTSTATUS",
  SET_PRODUCTSTATUSES = "SET_PRODUCTSTATUSES",
  CLEAR_PRODUCTSTATUSES = "CLEAR_PRODUCTSTATUSES",
}